# apps/core/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Sol Home Automation API
  version: "0.2.0"
  description: Backend API for the Sol home automation system.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/homes:
    get:
      tags: [Homes]
      summary: List homes for current user
      operationId: listHomes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Cursor-paginated homes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeCursorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Homes]
      summary: Create home
      operationId: createHome
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHomeRequest"
      responses:
        "201":
          description: Home created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Home"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/homes/{id}:
    parameters:
      - $ref: "#/components/parameters/HomeId"
    get:
      tags: [Homes]
      summary: Get home by ID
      operationId: getHome
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Home details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Home"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Homes]
      summary: Update home
      operationId: updateHome
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHomeRequest"
      responses:
        "200":
          description: Home updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Home"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Homes]
      summary: Delete home
      operationId: deleteHome
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Home deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/homes/{id}/members/{userId}/role:
    parameters:
      - $ref: "#/components/parameters/HomeId"
      - $ref: "#/components/parameters/UserId"
    patch:
      tags: [Members]
      summary: Update member role
      operationId: updateMemberRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRoleRequest"
      responses:
        "204":
          description: Role updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/homes/{id}/invitations:
    parameters:
      - $ref: "#/components/parameters/HomeId"
    get:
      tags: [Invitations]
      summary: List invitations for home
      operationId: listHomeInvitations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/InvitationStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Cursor-paginated invitations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeInvitationCursorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Invitations]
      summary: Invite user by email
      operationId: inviteByEmail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteByEmailRequest"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeInvitation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/homes/{id}/invitations/{invId}:
    parameters:
      - $ref: "#/components/parameters/HomeId"
      - $ref: "#/components/parameters/InvitationId"
    delete:
      tags: [Invitations]
      summary: Cancel invitation
      operationId: cancelInvitation
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Invitation cancelled
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/invitations/{token}/accept:
    parameters:
      - $ref: "#/components/parameters/InvitationToken"
    post:
      tags: [Invitations]
      summary: Accept invitation
      operationId: acceptInvitation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Home joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Home"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
    HomeId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
    InvitationId:
      name: invId
      in: path
      required: true
      schema:
        type: string
    InvitationToken:
      name: token
      in: path
      required: true
      schema:
        type: string
...
    InvitationStatus:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/InvitationStatus"
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict in current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CursorMeta:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: ["string", "null"]
        has_more:
          type: boolean
...
    MemberRole:
      type: string
      enum: [owner, admin, member]
    InvitationStatus:
      type: string
      enum: [pending, accepted, declined, expired]
    Home:
      type: object
      required: [id, name, owner_id, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        owner_id:
          type: string
        my_role:
          $ref: "#/components/schemas/MemberRole"
        member_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    HomeCursorResponse:
      allOf:
        - $ref: "#/components/schemas/CursorMeta"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Home"
    CreateHomeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateHomeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
...
    UpdateMemberRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, member]
    HomeInvitation:
      type: object
      required: [id, home_id, inviter_id, invitee_email, invitee_is_user, status, expires_at, created_at]
      properties:
        id:
          type: string
        home_id:
          type: string
        inviter_id:
          type: string
        invitee_email:
          type: string
          format: email
        invitee_is_user:
          type: boolean
        token:
          type: string
        status:
          $ref: "#/components/schemas/InvitationStatus"
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    HomeInvitationCursorResponse:
      allOf:
        - $ref: "#/components/schemas/CursorMeta"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/HomeInvitation"
...
    InviteByEmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...